+++ aiohttp/client.py
@@ -6,10 +6,10 @@ import hashlib
 import os
 import sys
 import traceback
-import urllib.parse
 import warnings
 
 from multidict import CIMultiDict, MultiDict, MultiDictProxy, istr
+from yarl import URL
 
 import aiohttp
 
@@ -180,8 +180,11 @@ class ClientSession:
             for i in skip_auto_headers:
                 skip_headers.add(istr(i))
 
+        if isinstance(proxy, str):
+            proxy = URL(proxy)
+
         while True:
-            url, _ = urllib.parse.urldefrag(url)
+            url = URL(url).with_fragment(None)
 
             cookies = self._cookie_jar.filter_cookies(url)
 
@@ -237,15 +240,15 @@ class ClientSession:
                     if headers.get(hdrs.CONTENT_LENGTH):
                         headers.pop(hdrs.CONTENT_LENGTH)
 
-                r_url = (resp.headers.get(hdrs.LOCATION) or
-                         resp.headers.get(hdrs.URI))
+                r_url = URL(resp.headers.get(hdrs.LOCATION) or
+                            resp.headers.get(hdrs.URI))
 
-                scheme = urllib.parse.urlsplit(r_url)[0]
+                scheme = r_url.scheme
                 if scheme not in ('http', 'https', ''):
                     resp.close()
                     raise ValueError('Can redirect only to http or https')
                 elif not scheme:
-                    r_url = urllib.parse.urljoin(url, r_url)
+                    r_url = url.join(r_url)
 
                 url = r_url
                 params = None