+++ aiohttp/connector.py
@@ -11,6 +11,8 @@ from itertools import chain
 from math import ceil
 from types import MappingProxyType
 
+from yarl import URL
+
 import aiohttp
 
 from . import hdrs, helpers
@@ -638,9 +640,7 @@ class TCPConnector(BaseConnector):
             raise ProxyConnectionError(*exc.args) from exc
 
         if not req.ssl:
-            req.path = '{scheme}://{host}{path}'.format(scheme=req.scheme,
-                                                        host=req.netloc,
-                                                        path=req.path)
+            req.path = str(req.url)
         if hdrs.AUTHORIZATION in proxy_req.headers:
             auth = proxy_req.headers[hdrs.AUTHORIZATION]
             del proxy_req.headers[hdrs.AUTHORIZATION]
@@ -723,6 +723,7 @@ class ProxyConnector(TCPConnector):
                          conn_timeout=conn_timeout,
                          keepalive_timeout=keepalive_timeout,
                          limit=limit, loop=loop)
+        assert isinstance(proxy, URL)
         self._proxy = proxy
         self._proxy_auth = proxy_auth