+++ aiohttp/web.py
@@ -16,7 +16,6 @@ from .web_reqrep import *  # noqa
 from .web_urldispatcher import *  # noqa
 from .web_ws import *  # noqa
 
-
 __all__ = (web_reqrep.__all__ +
            web_exceptions.__all__ +
            web_urldispatcher.__all__ +